pyproject.toml
[project]
name = "rsa-paper"
version = "0.1.0"
description = ""
authors = [
    {name = "bacor",email = "mail@bascornelissen.nl"}
]
readme = "README.md"
requires-python = "3.11.13"
dependencies = [
    "ipykernel (>=7.0.0,<8.0.0)",
    "jupyterlab (>=4.4.9,<5.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "scipy (>=1.16.2,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "rich (>=14.2.0,<15.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "jupytext (>=1.17.3,<2.0.0)",
    "networkx (>=3.5,<4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
"xlrd (>=2.0.2,<3.0.0)",
"rhythmic-segments (>=0.1.10,<0.2.0)",
"scikit-learn (>=1.7.2,<2.0.0)",
"mpltern (>=1.0.4,<2.0.0)",
"umap-learn (>=0.5.9.post2,<0.6.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
